[service] Move `serviceHost` to its own file
